# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for AI Prompt Shell",
        "title": "AI Prompt Shell API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/environs": {
            "get": {
                "description": "Get all defined environment variables in system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Environs"
                ],
                "summary": "List all environment variables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/environs/{environ_id}": {
            "get": {
                "description": "Get value of specified environment variable",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Environs"
                ],
                "summary": "Get environment variable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment variable ID",
                        "name": "environ_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/api/extensions": {
            "get": {
                "description": "Get all available prompt extension IDs in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Extensions"
                ],
                "summary": "List all prompt extension IDs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/extensions/{extension_id}": {
            "get": {
                "description": "Get detailed information of prompt extension by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Extensions"
                ],
                "summary": "Get specified prompt extension details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Extension ID",
                        "name": "extension_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dao.PromptExtension"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/prompts": {
            "get": {
                "description": "Get all available prompt templates in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "List all prompt templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/prompts/{prompt_id}": {
            "get": {
                "description": "Get detailed information of prompt template by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "Get specified prompt template details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt template ID",
                        "name": "prompt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dao.Prompt"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/prompts/{prompt_id}/chat": {
            "post": {
                "description": "Chat interaction with LLM using specified prompt template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "Interact with LLM using prompt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt template ID",
                        "name": "prompt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chat parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/prompts/{prompt_id}/render": {
            "post": {
                "description": "Render the prompt template with given args",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "Render specified prompt template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt template ID",
                        "name": "prompt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rendering parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tools": {
            "get": {
                "description": "Get a list of all available tools in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "List all tools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tools/{tool_id}": {
            "get": {
                "description": "Get detailed information about specified tool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Get tool details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "工具ID",
                        "name": "tool_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dao.Tool"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ChatModelRequest": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string"
                },
                "args": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "dao.Contributes": {
            "type": "object",
            "properties": {
                "dependences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.Dependence"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "prompts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.Prompt"
                    }
                }
            }
        },
        "dao.Dependence": {
            "type": "object",
            "properties": {
                "failStrategy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dao.Engines": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dao.Grpc": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dao.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "dao.Prompt": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.Message"
                    }
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "prompt": {
                    "type": "string"
                },
                "returns": {
                    "type": "object",
                    "additionalProperties": true
                },
                "supports": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dao.PromptExtension": {
            "type": "object",
            "properties": {
                "contributes": {
                    "$ref": "#/definitions/dao.Contributes"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "engines": {
                    "$ref": "#/definitions/dao.Engines"
                },
                "extensionType": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dao.Restful": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dao.Tool": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "examples": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "grpc": {
                    "$ref": "#/definitions/dao.Grpc"
                },
                "module": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "restful": {
                    "$ref": "#/definitions/dao.Restful"
                },
                "returns": {
                    "type": "object",
                    "additionalProperties": true
                },
                "supports": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "service.ChatResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "finish_reason": {
                                "type": "string"
                            },
                            "index": {
                                "type": "integer"
                            },
                            "logprobs": {
                                "type": "object"
                            },
                            "message": {
                                "type": "object",
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "properties": {
                        "completion_tokens": {
                            "type": "integer"
                        },
                        "prompt_tokens": {
                            "type": "integer"
                        },
                        "total_tokens": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}